backend/requirements.txt
fastapi
uvicorn
python-pam
PyJWT
python-multipart
aiofiles
asyncssh
aioftp
aiosqlite
aiohttp
